test: gather the shared test doubles under tests/Support
Four Builds* traits still lived next to whichever test happened to need them
first, which meant a consumer in another namespace had to reach across the tree
to find them. They have 3 to 4 consumers each now, so they belong in
tests/Support alongside the doubles introduced earlier in this branch.

DataTablesBundleTest, MutationServiceWiringTest and DataTablesExtensionTest all
booted TwigAppKernel with byte-identical setup. The only apparent difference was
that the Twig extension test also held test.service_container in a property,
which is the same booted kernel seen through a second handle, not a different
configuration. So BootsTwigKernel declares both handles and needs no flag and no
argument to serve all three.

Constraint: this pass moves code and changes nothing else. Assertion count was
  identical before and after (905 tests, 2950 assertions on both runs); the
  900/2938 reported later is the AsDataTableTest fold landing in the same tree.
Rejected: a trait taking a config array or a boolean to cover all three kernel
  consumers. A double that needs a flag to serve its callers is worse than the
  duplication it replaces.
Rejected: making the two remaining consumers adopt the trait's service()
  accessor. It would add assertIsObject calls, and this pass must not move the
  assertion count.
Confidence: high. Purely mechanical; git detected all four moves as renames.
Scope-risk: low. Tests only, no file under src/ touched.
Directive: user asked to continue through the remaining batches to completion.
Tested: composer fix (0 of 370 files changed), vendor/bin/phpunit
  (900 tests, 2938 assertions, green), git diff --name-only -- src empty.
Not-tested: MutationServiceWiringTest::readPrivateProperty stayed local; it has
  a single consumer and sharing it was not in scope.
